+++ src/keyboard/SDLUSLayout.h
@@ -28,36 +28,38 @@ public:
     layoutName = "US";
     addCommonKeys();
 
-    mapping[{SDLK_MINUS, false, false}] = C64_KEYCODE_MINUS;
-    mapping[{SDLK_PERIOD, false, false}] = C64_KEYCODE_PERIOD;
-    mapping[{SDLK_COMMA, false, false}] = C64_KEYCODE_COMMA;
-    mapping[{SDLK_4, true, false}] = C64_KEYCODE_DOLLAR;
-    mapping[{SDLK_COMMA, true, false}] = C64_KEYCODE_LESS;
-    mapping[{SDLK_QUOTE, false, false}] = C64_KEYCODE_QUOTE;
+    mapping[{(uint32_t)SDLK_MINUS, false, false}] = C64_KEYCODE_MINUS;
+    mapping[{(uint32_t)SDLK_PERIOD, false, false}] = C64_KEYCODE_PERIOD;
+    mapping[{(uint32_t)SDLK_COMMA, false, false}] = C64_KEYCODE_COMMA;
+    mapping[{(uint32_t)SDLK_4, true, false}] = C64_KEYCODE_DOLLAR;
+    mapping[{(uint32_t)SDLK_COMMA, true, false}] = C64_KEYCODE_LESS;
+    mapping[{(uint32_t)SDLK_QUOTE, false, false}] = C64_KEYCODE_QUOTE;
 
-    mapping[{SDLK_PERIOD, true, false}] = C64_KEYCODE_GREATER;
-    mapping[{SDLK_SLASH, true, false}] = C64_KEYCODE_QUESTION;
-    mapping[{SDLK_EQUALS, true, false}] = C64_KEYCODE_PLUS;
-    mapping[{SDLK_8, true, false}] = C64_KEYCODE_ASTERISK;
-    mapping[{SDLK_SLASH, false, false}] = C64_KEYCODE_SLASH;
-    mapping[{SDLK_EQUALS, false, false}] = C64_KEYCODE_EQUALS;
-    mapping[{SDLK_SEMICOLON, true, false}] = C64_KEYCODE_COLON;
-    mapping[{SDLK_SEMICOLON, false, false}] = C64_KEYCODE_SEMICOLON;
-    mapping[{SDLK_MINUS, true, false}] = C64_KEYCODE_POUND;
-    mapping[{SDLK_QUOTE, true, false}] = C64_KEYCODE_QUOTEDBL;
-    mapping[{SDLK_5, true, false}] = C64_KEYCODE_PERCENT;
-    mapping[{SDLK_7, true, false}] = C64_KEYCODE_AMPERSAND;
-    mapping[{SDLK_9, true, false}] = C64_KEYCODE_LEFTPAREN;
-    mapping[{SDLK_0, true, false}] = C64_KEYCODE_RIGHTPAREN;
+    mapping[{(uint32_t)SDLK_PERIOD, true, false}] = C64_KEYCODE_GREATER;
+    mapping[{(uint32_t)SDLK_SLASH, true, false}] = C64_KEYCODE_QUESTION;
+    mapping[{(uint32_t)SDLK_EQUALS, true, false}] = C64_KEYCODE_PLUS;
+    mapping[{(uint32_t)SDLK_8, true, false}] = C64_KEYCODE_ASTERISK;
+    mapping[{(uint32_t)SDLK_SLASH, false, false}] = C64_KEYCODE_SLASH;
+    mapping[{(uint32_t)SDLK_EQUALS, false, false}] = C64_KEYCODE_EQUALS;
+    mapping[{(uint32_t)SDLK_SEMICOLON, true, false}] = C64_KEYCODE_COLON;
+    mapping[{(uint32_t)SDLK_SEMICOLON, false, false}] = C64_KEYCODE_SEMICOLON;
+    mapping[{(uint32_t)SDLK_MINUS, true, false}] = C64_KEYCODE_POUND;
+    mapping[{(uint32_t)SDLK_QUOTE, true, false}] = C64_KEYCODE_QUOTEDBL;
+    mapping[{(uint32_t)SDLK_5, true, false}] = C64_KEYCODE_PERCENT;
+    mapping[{(uint32_t)SDLK_7, true, false}] = C64_KEYCODE_AMPERSAND;
+    mapping[{(uint32_t)SDLK_9, true, false}] = C64_KEYCODE_LEFTPAREN;
+    mapping[{(uint32_t)SDLK_0, true, false}] = C64_KEYCODE_RIGHTPAREN;
 
-    mapping[{SDLK_2, true, false}] = C64_KEYCODE_AT;
-    mapping[{SDLK_3, true, false}] = C64_KEYCODE_HASH;
+    mapping[{(uint32_t)SDLK_2, true, false}] = C64_KEYCODE_AT;
+    mapping[{(uint32_t)SDLK_3, true, false}] = C64_KEYCODE_HASH;
 
-    mapping[{SDLK_1, true, false}] = C64_KEYCODE_EXCLAIM;
-    mapping[{SDLK_6, true, false}] = C64_KEYCODE_UPARROW;
-    mapping[{SDLK_BACKQUOTE, false, false}] = C64_KEYCODE_LEFTARROW;
-    mapping[{SDLK_LEFTBRACKET, false, false}] = C64_KEYCODE_LEFTBRACKET;
-    mapping[{SDLK_RIGHTBRACKET, false, false}] = C64_KEYCODE_RIGHTBRACKET;
+    mapping[{(uint32_t)SDLK_1, true, false}] = C64_KEYCODE_EXCLAIM;
+    mapping[{(uint32_t)SDLK_6, true, false}] = C64_KEYCODE_UPARROW;
+    mapping[{(uint32_t)SDLK_BACKQUOTE, false, false}] = C64_KEYCODE_LEFTARROW;
+    mapping[{(uint32_t)SDLK_LEFTBRACKET, false, false}] =
+        C64_KEYCODE_LEFTBRACKET;
+    mapping[{(uint32_t)SDLK_RIGHTBRACKET, false, false}] =
+        C64_KEYCODE_RIGHTBRACKET;
   }
 };
